Add files via upload
This is an admin dashboard for the YobYong's ordering system. Specifically for the admin and staff to see their restaurant's sales by each month. Objective of this dashboard is making it easy for admin and staff to see total sale through online ordering. They can also add new menu and update the status of each depending on the stock for the month.